requirements.txt
annotated-types==0.5.0
anyio==3.7.1
azure-core==1.30.2
azure-functions==1.20.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.30.0
msal-extensions==1.2.0
orjson==3.9.7
portalocker==2.10.1
pycparser==2.22
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
pywin32==306
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.2.2
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
